#26 Refactor: minor fixes, work with timezone
