+++ backend/flower/core/unit_utils.py
@@ -1,7 +1,10 @@
 import os
-import base64
+import datetime
 
+from pytz import utc
 from uuid import uuid4
+from base64 import b64decode
+
 from .. import config
 
 
@@ -9,7 +12,7 @@ class MediaUtils:
     @staticmethod
     async def save_file_base64(ext_file):
         service_data, base64_data = ext_file.split(',')
-        ext = service_data.split(';')[0][5:].split('/')[1]
+        ext = service_data.split(';')[0].split('/')[-1]
 
         file_name = MediaUtils.create_filename(ext)
 
@@ -22,7 +25,7 @@ class MediaUtils:
         path = os.path.join(path, file_name[2:])
 
         with open(os.path.join(config.MEDIA_FOLDER, path), 'wb') as fh:
-            fh.write(base64.b64decode(base64_data))
+            fh.write(b64decode(base64_data))
 
         return path
 
@@ -53,3 +56,15 @@ class MediaUtils:
     def generate_full_path(path):
         # PS: normpath на Windows преобразует обычные слеши в обратные
         return os.path.normpath(os.path.join(config.MEDIA_FOLDER, path))
+
+
+def convert_to_utc(dt):
+    """Return same datetime if it's aware or sets it's timezone to UTC."""
+
+    if dt is None:
+        dt = datetime.datetime.utcfromtimestamp(0)
+
+    if dt.tzinfo is None or dt.tzinfo.utcoffset(dt) is None:
+        return dt.replace(tzinfo=utc)
+
+    return dt
